game-api/requirements.txt
alembic==1.9.2
bidict==0.22.1
click==8.1.3
dnspython==2.3.0
eventlet==0.33.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==23.0
PyJWT==2.6.0
python-dotenv==0.21.1
python-engineio==4.3.4
python-socketio==5.7.2
simple-websocket==0.9.0
six==1.16.0
SQLAlchemy==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
wsproto==1.2.0
zope.event==4.6
zope.interface==5.5.2
